+++ tucans-planmatica/.env
@@ -1,0 +1,2 @@
+GOOGLE_ID=240617596486-e60g1iuj42ff45sevfsdrhntc5jmm2e2.apps.googleusercontent.com
+GOOGLE_SECRET=lyP4E5oAinRoOjckt0Fgfh9K
