+++ tucans-planmatica/components/MainHead.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Head from "next/head";
-
+import icon from "../public/favicon.ico";
 export default function MainHead(props) {
     return (
         <div>
@@ -9,7 +9,7 @@ export default function MainHead(props) {
                 <meta http-equiv="X-UA-Compatible" content="IE=edge" />
                 <meta name="viewport" content="width=device-width, initial-scale=1.0" />
                 <title> {props.tituloPestana} </title>
-                <link rel="shortcut icon" href={"../src/icon.png"} />
+                <link rel="shortcut icon" href={props.icon} />
                 <link rel="preconnect" href="https://fonts.googleapis.com" />
                 <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
                 <link href="https://fonts.googleapis.com/css2?family=Work+Sans:ital,wght@0,300;1,300&display=swap" rel="stylesheet"></link>
