+++ tucans-planmatica/pages/_app.js
@@ -1,7 +1,10 @@
-import '../styles/globals.css'
+import { Provider } from "next-auth/client";
+import '../styles/globals.css';
 
 function MyApp({ Component, pageProps }) {
-  return <Component {...pageProps} />
+  return <Provider session={pageProps.session}>
+    <Component {...pageProps} />
+  </Provider>
 }
 
 export default MyApp
