+++ tucans-planmatica/pages/session/IniciarSesion.jsx
@@ -1,11 +1,15 @@
 import React from 'react';
-import styles from "../styles/Forms.module.css";
+import LayoutIndex from '../../components/LayoutIndex';
+import MainHead from '../../components/MainHead';
+import styles from "../../styles/Forms.module.css";
 
 export default function IniciarSesion() {
     return (
         <div>
+            <MainHead title="Iniciar Sesion" />
+            <LayoutIndex>
             <form className={styles.root} >
-                <h2 className={styles.title} >Iniciar Sesion</h2>
+                <h2 className={styles.title} ><a href="/">←</a>Iniciar Sesion</h2>
 
                 <div className={styles.cont_input}>
                     <label className={styles.text} >Nombre de usuario o correo</label>
@@ -18,6 +22,7 @@ export default function IniciarSesion() {
                 <button className={styles.btnSubmit} type="submit">Iniciar Sesion</button>
 
             </form>
+            </LayoutIndex>
         </div>
     )
 }
