+++ tucans-planmatica/styles/Home.module.css
@@ -74,6 +74,7 @@
   background-repeat: no-repeat;
   position: relative;
   width: 100%;
+  height: 120px;
   display: inline-block;
   text-align: center;
   align-items: center;
@@ -136,14 +137,82 @@
 }
 
 .container_0 {
+  background-image: linear-gradient(to top, rgba(240, 240, 240, 0.8) 0%, rgba(255, 255, 255, 0.825) 100%), url(../src/head_bg.jpg);
+  background-size: cover;
+  background-attachment: fixed;
+  background-position: center;
+  background-repeat: no-repeat;
   display: inline-flexbox;
 }
 
 .container_1 {
   width: 90%;
   margin: auto;
-  box-shadow: 0px 0px 7px rgb(50, 50, 50);
+  text-align: center;
 }
+
+.container_2 {
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+}
+
+.container_2 table {
+  width: 100%;
+  text-align: center;
+}
+
+.container_3 {
+  width: 85%;
+  text-align: center;
+}
+
+.container_4 {
+  width: 50%;
+  min-width: 350px;
+  display: flex;
+  justify-content: space-between;
+}
+
+.btnSubmit {
+  margin: 16px;
+  margin-top: 15.8vh;
+  border-radius: 4px;
+  height: 35px;
+  font-size: 18px;
+  color: white;
+  background-color: var(--main-color);
+  border: 0;
+  transition: 0.3s;
+}
+
+.btnSubmit:hover {
+  cursor: pointer;
+  background-color: var(--hover-color);
+  transform: scale(1.1, 1.1);
+}
+
+.aside_0 {
+  width: 20%;
+  border: 1px solid #000000;
+  text-align: center;
+}
+
+/*Contenido*/
+
+.cont_img {
+  width: fit-content;
+  border-radius: 10px;
+  box-shadow: 0px 0px 10px rgb(50, 50, 50);
+}
+
+.index_img {
+  border-radius: 10px;
+  padding-left: 10px;
+  transition: 0.3s;
+}
+
+/*Footer*/
 
 .footy {
   background-image: linear-gradient(to top, rgba(46, 33, 23, 0.8) 0%, rgba(48, 31, 18, 0.8) 100%), url(../src/head_bg.jpg);
@@ -151,7 +220,7 @@
   background-attachment: fixed;
   background-position: center;
   background-repeat: no-repeat;
-  position: absolute;
+  margin-top: 32px;
   bottom: 0;
   width: 100%;
   box-shadow: 0px 0px 7px rgb(50, 50, 50);
@@ -183,7 +252,7 @@
 .footy a {
   text-decoration: none;
   color: rgb(250, 250, 250);
-  transition: 0.175s ease-in-out;
+  transition: 0.2s ease-in-out;
 }
 
 .footy a:hover {
